refactor: clean arch usecases pattern
