+++ server/src/main/java/com/mvgm/snug_server/core/usecases/BuildEmail.java
