+++ server/src/main/java/com/mvgm/snug_server/core/usecases/GenerateJwtToken.java
